refactor: renamed model_analyzer to analyze
