+++ lightwood/analysis/__init__.py
@@ -1,4 +1,4 @@
-from lightwood.analysis.model_analyzer import model_analyzer
+from lightwood.analysis.analyze import model_analyzer
 from lightwood.analysis.explain import explain
 
 __all__ = ['model_analyzer', 'explain']
